Fix Tests for Config
